requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0